split off bespoke functions from vendor script (#646)
