+++ blocks/magazine-articles/magazine-articles.js
@@ -4,9 +4,11 @@ import {
 } from '../../scripts/common.js';
 import {
   ffetch,
+} from '../../scripts/lib-ffetch.js';
+import {
   createList,
   splitTags,
-} from '../../scripts/lib-ffetch.js';
+} from '../../scripts/magazine-press.js';
 import {
   createOptimizedPicture,
   getMetadata,
